requirements.txt
maritalk
python-dotenv
chainlit
openai